+++ src/main/java/com/gmbdesign/stratio/stratiowars/presentation/rest/RebelDecryptorResource.java
@@ -2,7 +2,7 @@ package com.gmbdesign.stratio.stratiowars.presentation.rest;
 
 import com.gmbdesign.stratio.stratiowars.presentation.dto.DecryptedCoordinateListDTO;
 import com.gmbdesign.stratio.stratiowars.presentation.dto.EncryptedCoordinateListDTO;
-import com.gmbdesign.stratio.stratiowars.test.service.IDecryptorService;
+import com.gmbdesign.stratio.stratiowars.service.IDecryptorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -13,12 +13,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class RebelDecryptorResource {
 
     @Autowired
-    IDecryptorService decryptorService;
+    private IDecryptorService decryptorService;
 
     @RequestMapping(method = RequestMethod.GET, produces = "application/json")
     public DecryptedCoordinateListDTO getCoordinatesDecrypted (EncryptedCoordinateListDTO encryptedCoordinateListDTO) {
-        DecryptedCoordinateListDTO decryptedCoordinateListDTO = null;
-        decryptedCoordinateListDTO = decryptorService.decryptCoordinateList(encryptedCoordinateListDTO);
-        return decryptedCoordinateListDTO;
+        return decryptorService.decryptCoordinateList(encryptedCoordinateListDTO);
     }
 }
