+++ src/main/java/com/gmbdesign/stratio/stratiowars/service/IDecryptorService.java
@@ -1,4 +1,4 @@
-package com.gmbdesign.stratio.stratiowars.test.service;
+package com.gmbdesign.stratio.stratiowars.service;
 
 import com.gmbdesign.stratio.stratiowars.presentation.dto.DecryptedCoordinateListDTO;
 import com.gmbdesign.stratio.stratiowars.presentation.dto.EncryptedCoordinateListDTO;
@@ -8,8 +8,8 @@ public interface IDecryptorService {
     /**
      * Decrypt a list of encrypted coordinates of Death Star position. May the Force be with you.
      *
-     * @param encryptedCoordinateListDTO
-     * @return
+     * @param encryptedCoordinateListDTO encrypted coordinate list
+     * @return decrypted coordinate list
      */
     DecryptedCoordinateListDTO decryptCoordinateList(EncryptedCoordinateListDTO encryptedCoordinateListDTO);
 }
